refactor(web): get game page dynamically
